Refactor RSC setup helpers
